refactor commons into common header
