+++ README.md
@@ -15,7 +15,8 @@ Thanks to [GenosseFlosse](https://github.com/GenosseFlosse) for providing help w
 1. Download or clone the HiGHS code. The *highsmex.cpp* file needs to be compiled with C++20 switch, hence it would be advisable to compile HiGHS with C++20 switch. To do so modify the CMakeLists.txt of HiGHS by changing the line 29 ```set(CMAKE_CXX_STANDARD 11)``` to ```set(CMAKE_CXX_STANDARD 20)``` and line 301 ```set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")``` to ```set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")```. Install HiGHS as described [here.](https://github.com/ERGO-Code/HiGHS/tree/master/cmake) This should create a static library named *highs* e.g., *highs.lib* on Windows.\
 **Note for macOS users** Your cmake command should look like the following \
 `cmake -S <path-to-source> -B <path-to-build> -DBUILD_SHARED_LIBS=OFF -DZLIB=OFF -DCMAKE_OSX_DEPLOYMENT_TARGET=13.4`
-2. Open script named *make_highsmex.m* in MATLAB and specify the inputs. Then execute the script. This builds both mex files of the project, named highsmex.mex* and highsmex_iis.mex* e.g., highsmex.mexw64 and highsmex_iis.mexw64 on 64-bit Windows. The two are independent of each other, so a single one can be built by naming its source file, e.g. ```make_highsmex('1.15.1', 'highsmex.cpp')```.
+2. Open script named *make_highsmex.m* in MATLAB and specify the inputs. Then execute the script. This builds both mex files of the project, named highsmex.mex* and highsmex_iis.mex* e.g., highsmex.mexw64 and highsmex_iis.mexw64 on 64-bit Windows. The two are independent of each other, so a single one can be built by naming its source file, e.g. ```make_highsmex('1.15.1', 'highsmex.cpp')```.\
+The code shared by the two, i.e. the MATLAB type helpers, the conversion of the HiGHS options and info structs, the matrix conversion, and the mex logging and error handling, lives in the header *highsmex_common.hpp*. It is header-only, so no extra source file has to be passed to *mex*.
 
 
 ## Documentation
